% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_example-data.R
\docType{data}
\name{sp_bgsmtr_example_data}
\alias{sp_bgsmtr_example_data}
\title{Example Structural Neuroimaging and Genetic Data for Spatial Model.}
\format{A list with four components: "SNP_data", "SNP_groups", "BrainMeasures", "path_data"
SNP_data is a 486-by-632 matrix containing minor allele counts for 632 subjects and 486 SNPs.
neighbourhood_structure is a 12 by 12 first order neighbourhood structure matrix.
BrainMeasures is a 24-by-632 matrix containing simulated volumetric and cortical thickness measures for 24 regions of interest.

path_data is a list of two elements. The fist element is "lambda_v", which is a vector of different lambda squared values. The second element is "W_est_list", which is a list containing estimated coefficients matrices W for each corresponding lambda squared value.}
\usage{
data(sp_bgsmtr_example_data)
}
\description{
This example dataset has simulated dataset with 632 subjects, 486 SNPs from 24 structural neuroimaging measures. It also contains example datasets used for regularization path plotting function.
}
\examples{
data(sp_bgsmtr_example_data)
names(sp_bgsmtr_example_data)
dim(sp_bgsmtr_example_data$SNP_data)
dim(sp_bgsmtr_example_data$BrainMeasures)
dim(sp_bgsmtr_example_data$neighbourhood_structure)
length(sp_bgsmtr_example_data$path_data$lambda_v)
length(sp_bgsmtr_example_data$path_data$W_est_list)
}
\keyword{datasets}
